判断缩放与旋转手势

+++ chapter11/src/main/java/com/jpc/chapter11/utils/Utils.kt
@@ -8,10 +8,10 @@ import android.widget.LinearLayout
 
 object Utils {
     // 根据手机的分辨率从 dp 的单位 转成为 px(像素)
-    fun dip2px(context: Context, dpValue: Float): Int {
+    fun dip2px(context: Context, dpValue: Float): Float {
         // 获取当前手机的像素密度
         val scale = context.resources.displayMetrics.density
-        return (dpValue * scale + 0.5f).toInt() // 四舍五入取整
+        return (dpValue * scale + 0.5f).toInt().toFloat() // 四舍五入取整
     }
 
     // 根据手机的分辨率从 px(像素) 的单位 转成为 dp
